RWlockCommon.sh: Update locking APIs and extract common code
LockAnalysis.sh and RWlockAnalysis.sh duplicate the reader-writer
locking analysis code, so this commit extracts it into a separate
RWlockCommon.sh file.

While in the area, also update the locking APIs, including the
long-overdue addition of reader-writer semaphores.  These were originally
omitted because the original RCU API did not permit blocking, and was
thus more comparable to reader-writer spinlocks than to reader-writer
semaphores.  But the advent of SRCU changed that, and these scripts are
belatedly taking this into account.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>